A few upload variables that didn't make it last time.
A few more words of justification for this change. I had a thought that we could just leave upload enabled for existing nodes and only add filefield for where it's really needed such as a podcasting feature but... I think it'll be simpler if we have a uniform enviornment and the one thing we're losing, comment uploads will be restored once we can upgrade to Drupal 7 which hopefully come within 6-7 months.